+++ iperf3/iperf3.py
@@ -31,7 +31,7 @@ except ImportError:
     from Queue import Queue  # Python2 compatibility
 
 
-__version__ = '0.1.8'
+__version__ = '0.1.9'
 
 
 MAX_UDP_BULKSIZE = (65535 - 8 - 20)
@@ -802,24 +802,33 @@ class TestResult(object):
 
             # TCP specific test results
             if self.protocol == 'TCP':
-                self.sent_bytes = self.json['end']['sum_sent']['bytes']
-                self.sent_bps = self.json['end']['sum_sent']['bits_per_second']
-                self.received_bytes = self.json['end']['sum_received']['bytes']
-                self.received_bps = self.json['end']['sum_received']['bits_per_second']
+                sent_json = self.json['end']['sum_sent']
+                self.sent_bytes = sent_json['bytes']
+                self.sent_bps = sent_json['bits_per_second']
 
-                #  Bits are measured in 10**3 terms, Bytes are measured in 2**10 terms
-                self.sent_kbps = self.sent_bps / 1000           # Kilobits per second
-                self.sent_Mbps = self.sent_kbps / 1000          # Megabits per second
-                self.sent_kB_s = self.sent_bps / (8 * 1024)     # kiloBytes per second
-                self.sent_MB_s = self.sent_kB_s / 1024          # MegaBytes per second
+                recv_json = self.json['end']['sum_received']
+                self.received_bytes = recv_json['bytes']
+                self.received_bps = recv_json['bits_per_second']
 
-                self.received_kbps = self.received_bps / 1000   # Kilobits per second
-                self.received_Mbps = self.received_kbps / 1000  # Megabits per second
-                self.received_kB_s = self.received_bps / (8* 1024)     # kiloBytes per second
-                self.received_MB_s = self.received_kB_s / 1024     # MegaBytes per second
+                # Bits are measured in 10**3 terms
+                # Bytes are measured in 2**10 terms
+                # kbps = Kilobits per second
+                # Mbps = Megabits per second
+                # kB_s = kiloBytes per second
+                # MB_s = MegaBytes per second
+
+                self.sent_kbps = self.sent_bps / 1000
+                self.sent_Mbps = self.sent_kbps / 1000
+                self.sent_kB_s = self.sent_bps / (8 * 1024)
+                self.sent_MB_s = self.sent_kB_s / 1024
+
+                self.received_kbps = self.received_bps / 1000
+                self.received_Mbps = self.received_kbps / 1000
+                self.received_kB_s = self.received_bps / (8 * 1024)
+                self.received_MB_s = self.received_kB_s / 1024
 
                 # retransmits only returned from client
-                self.retransmits = self.json['end']['sum_sent'].get('retransmits')
+                self.retransmits = sent_json.get('retransmits')
 
             # UDP specific test results
             elif self.protocol == 'UDP':